refactor: 문서 최신 버전 조회 로직 및 버전 타입 변경
- 문서 버전 필드를 Long에서 Integer로 변경
- DocumentGetResponse 프로젝션에 값 객체 내부 필드 사용
- DocumentRepositoryImpl 최신 히스토리 조회 쿼리 단순화
- 최신 버전 조회 시 예외 처리를 서비스 계층으로 이동
- DocumentHistory 엔티티에 is_deleted 컬럼 추가

+++ server/src/main/java/com/openmpy/server/document/dto/response/DocumentGetResponse.java
@@ -12,14 +12,16 @@ public record DocumentGetResponse(
     DocumentCategory category,
     String author,
     String content,
-    Long version,
+    Integer version,
     Long size,
     LocalDateTime createdAt,
     LocalDateTime lastModifiedAt
 ) {
 
-    public static DocumentGetResponse from(final Document document,
-        final DocumentHistory documentHistory) {
+    public static DocumentGetResponse from(
+        final Document document,
+        final DocumentHistory documentHistory
+    ) {
         return new DocumentGetResponse(
             document.getId(),
             documentHistory.getId(),
